requirements.txt

flask
python-dotenv
whisper
torch
google-generativeai
gTTS
werkzeug
